// src/modules/admin/token-sale-rounds/dto/token-sale-overview.dto.ts
export class TokenSaleOverviewDto {
  total_rounds: number;
  total_mmp_sold_all_rounds: number;
  total_mpb_sold_all_rounds: number;
  total_mmp_referral_rewards_all_rounds: number;
  total_mpb_referral_rewards_all_rounds: number;
  total_mmp_net_sold_all_rounds: number;
  total_mpb_net_sold_all_rounds: number;
  total_swap_orders_all_rounds: number;
  total_referral_rewards_all_rounds: number;
  last_updated: Date;
} 